- Implement the vanity service on the frontend site for short links and include some plausibility checks before sending the request to the backend.

+++ src/app/components/common/home/home.controller.js
@@ -1,14 +1,14 @@
-/*
- * Copyright (C) Schweizerische Bundesbahnen SBB, 2016.
- */
 let self;
 class HomeController {
     constructor($http) {
         this.backendServiceUrl = 'http://localhost:8080/rest/1.0/shortlink/create?dest=';
+        this.backendVanityServiceUrl = 'http://localhost:8080/rest/1.0/shortlink/createVanityUrl?';
         this.baseDeleteUrl = 'http://localhost:8080/rest/1.0/shortlink/delete?id=';
         this.baseRedirectUrl = 'http://localhost:8080/x/';
         this.urlToCreate = '';
         this.urlToDelete = '';
+        this.txtIdForUrl = '';
+        this.urlId = '';
 
         this.title = 'URL Shortener Service';
         this.labelCreateButton = 'Generiere Shortlink';
@@ -20,14 +20,20 @@ class HomeController {
     }
 
     createUrl() {
+        if (this.urlToCreate === undefined || this.urlToCreate === '') return;
+
         this.http.post(this.backendServiceUrl + this.urlToCreate)
             .success(function (data) {
-                self.id = data.id;
+                self.urlId = data.id;
             });
     }
 
     deleteShortlink() {
+        if (this.urlToDelete === undefined || this.urlToDelete === '') return;
+
         var indexOfSlash = this.urlToDelete.indexOf('x/');
+        if (indexOfSlash == -1) return;
+
         var id = this.urlToDelete.substr(indexOfSlash + 2);
 
         this.http.delete(this.baseDeleteUrl + id).success(() => {
@@ -35,6 +41,23 @@ class HomeController {
         }).error(() => {
             self.deleted = false;
         });
+    }
+
+    createVanityUrl() {
+        if (this.txtIdForUrl === '' || this.urlToCreate === undefined || this.urlToCreate === '') return;
+
+        var params = 'text=' + this.txtIdForUrl + '&url=' + this.urlToCreate;
+        this.http.post(this.backendVanityServiceUrl + params)
+            .success(() => {
+                self.vanityUrlCreated = true;
+            })
+            .error(function (data) {
+                self.vanityUrlCreated = false;
+                if (data.status == 409) {
+                    self.vanityUrlErrorMessage = 'Der Text wird bereits gebraucht, bitte wählen ' +
+                        'Sie einen anderen Text.';
+                }
+            });
     }
 }
